Move scroll widget functions
